requirements.txt
# requirements.txt
newspaper3k
boundaryml
